+++ Kernel/Output/HTML/Layout.pm
@@ -4655,7 +4655,6 @@ sub _BuildSelectionDataRefCreate {
                             Value    => $ElementLongName,
                             Disabled => 1,
                         };
-
                     }
                     $Parents .= $Element . '::';
                 }
@@ -4664,6 +4663,7 @@ sub _BuildSelectionDataRefCreate {
                 push @NewDataLocal, {
                     Key      => $ValueHash->{Key},
                     Value    => $ValueHash->{Value},
+                    Selected => $ValueHash->{Selected} ? 1 : 0,
                     Disabled => $ValueHash->{Disabled} ? 1 : 0,
                 };
             }